[WIP] refine code in graph proxy
