+++ research/gaia/gremlin/gremlin_core/src/structure/mod.rs
@@ -21,7 +21,7 @@ mod property;
 use crate::generated::gremlin as pb;
 use crate::structure::codec::ParseError;
 use crate::FromPb;
-pub use element::{Edge, Element, GraphElement, Label, Vertex, VertexOrEdge, ID, ID_BITS};
+pub use element::{Edge, Element, GraphElement, Label, LabelId, Vertex, VertexOrEdge, ID, ID_BITS};
 pub use filter::*;
 pub use graph::*;
 pub use property::{DefaultDetails, Details, DynDetails, PropId, PropKey, Token};
